+++ charmander-domain/charmander-domain-common/src/main/java/org/cm/domain/term/Term.java
@@ -1,9 +1,10 @@
-package org.cm.domain.common;
+package org.cm.domain.term;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Lob;
 import java.time.LocalDate;
+import org.cm.domain.common.BaseEntity;
 
 @Entity(name = "terms")
 public class Term extends BaseEntity {
